修改字体

+++ base/canvas/freetype/src/pcf/pcfutil.h
@@ -30,22 +30,22 @@ THE SOFTWARE.
 
 
 #include <ft2build.h>
-#include "d:\freetype\include\freetype\config\ftconfig.h"
+#include FT_CONFIG_CONFIG_H
 
 
 FT_BEGIN_HEADER
 
   FT_LOCAL( void )
   BitOrderInvert( unsigned char*  buf,
-                  int             nbytes );
+                  size_t          nbytes );
 
   FT_LOCAL( void )
   TwoByteSwap( unsigned char*  buf,
-               int             nbytes );
+               size_t          nbytes );
 
   FT_LOCAL( void )
   FourByteSwap( unsigned char*  buf,
-                int             nbytes );
+                size_t          nbytes );
 
 FT_END_HEADER
